removed delete if status done

+++ src/app/appointments/appointments.component.ts
@@ -96,16 +96,18 @@ export class AppointmentsComponent implements OnInit {
   }
 
 
-  onDelete(e:Event, id : number){
+  onDelete(e:Event, id : number,status : string){
     const event = (e.target as any).value
     console.log(id)
+    if(status != 'done'){
     this.patientService.deleteAppointment(id).subscribe({
       next:()=>{
       },
       error:(error)=>{
         console.log("delete patient visit api failed",error)
       }
-    })
+    })}
+   
   }
   // Function to load appointments from the API
   loadAppointments(phoneNumber: string): Observable<GetPatientVisitDto[]> {
